refactor tests #46
