+++ src/TypealizR.Tests/MethodBuilder.Tests.cs
@@ -29,116 +29,4 @@ public class MethodBuilder_Tests
 		actual.Should().Be(expected);
 	}
 
-	[Theory]
-	[InlineData(
-		"Hello {0}",
-		"object _0"
-	)]
-	[InlineData(
-		"Hello {name}",
-		"object name"
-	)]
-	[InlineData(
-		"Hello {0}, today is {1}",
-		"object _0", "object _1"
-	)]
-	[InlineData(
-		"Hello {name}, today is {date}",
-		"object name", "object date"
-	)]
-	[InlineData(
-		"{date} Hello {name}, today is {date}",
-		"object name", "object date"
-	)]
-	[InlineData(
-		"Hello {name:s}, today is {date:d}",
-		"string name", "DateOnly date"
-	)]
-	[InlineData(
-		"{i:int} {s:string}  {dt:DateTime}  {dto:DateTimeOffset}  {d:DateOnly}  {t:TimeOnly}  {0}		   {1}          {name}",
-		"int i", "string s", "DateTime dt", "DateTimeOffset dto", "DateOnly d", "TimeOnly t", "object _0", "object _1", "object name"
-	)]
-	[InlineData(
-		"{i:i}   {s:s}       {dt:dt}        {dto:dto}             {d:d}         {t:t}",
-		"int i", "string s", "DateTime dt", "DateTimeOffset dto", "DateOnly d", "TimeOnly t"
-	)]
-
-	public void Extracts_Parameters(string input, params string[] expected)
-	{
-		var sut = new MethodBuilder(input, input);
-		var method = sut.Build(targetType, new(new("Ressource1.resx", input, 42)));
-
-		var actual = method.Parameters.Select(x => $"{x.Type} {x.DisplayName}").ToArray();
-
-		actual.Should().BeEquivalentTo(expected);
-	}
-
-	[Theory]
-	[InlineData(
-		"Hello {0}",
-		"object _0"
-	)]
-	[InlineData(
-		"Hello {name}",
-		"object name"
-	)]
-	[InlineData(
-		"Hello {0}, today is {1}",
-		"object _0, object _1"
-	)]
-	[InlineData(
-		"Hello {name}, today is {date}",
-		"object name, object date"
-	)]
-	[InlineData(
-		"{date} Hello {name}, today is {date}",
-		"object date, object name"
-	)]
-	public void Declares_Parameters_In_Signature(string input, string expectedPartialSignature)
-	{
-		var sut = new MethodBuilder(input, input);
-		var method = sut.Build(targetType, new (new("Ressource1.resx", input, 42)));
-		var writer = new IStringLocalizerExtensionMethodWriter(method);
-
-		var actual = writer.ToCSharp();
-		var expected = $"(this IStringLocalizer<{targetType.FullName}> that, {expectedPartialSignature})";
-
-		actual.Should().Be(expected);
-	}
-
-	[Theory]
-	[InlineData(
-		"Hello {0}",
-		"_0"
-	)]
-	[InlineData(
-		"Hello {name}",
-		"name"
-	)]
-	[InlineData(
-		"Hello {0}, today is {1}",
-		"_0, _1"
-	)]
-	[InlineData(
-		"Hello {name}, today is {date}",
-		"name, date"
-	)]
-	[InlineData(
-		"{date} Hello {name}, today is {date}",
-		"date, name"
-	)]
-	[InlineData(
-		"{0} {i:i} {s:s} {dt:dt} {dto:dto} {d:d} {t:t} {1}",
-		"_0, i, s, dt, dto, d, t, _1"
-	)]
-	public void Passes_Parameters_In_Invocation(string input, string expected)
-	{
-		var sut = new ParameterBuilder(input);
-		var parameters = sut.Build(new(new("Ressource1.resx", input, 42)));
-
-		var actual = parameters.Select(x => x.DisplayName).ToCommaDelimited();
-
-		actual.Should().BeEquivalentTo(expected);
-	}
-
 }
