+++ src/TypealizR.Tests/ParameterModel.Tests.cs
@@ -9,7 +9,7 @@ using TypealizR.StringLocalizer;
 
 namespace TypealizR.Tests;
 
-public class ParameterModel_Tests
+public class Parameter_Tests
 {
 
 	[Theory]
@@ -47,7 +47,125 @@ public class ParameterModel_Tests
 
 			warnings.Should().BeEquivalentTo(new[] { DiagnosticsFactory.TR0004.Id.ToString() });
 		}
-	
+
+	}
+
+	[Theory]
+	[InlineData(
+	"Hello {0}",
+	"object _0"
+	)]
+
+	[InlineData(
+	"Hello {name}",
+	"object name"
+	)]
+
+	[InlineData(
+	"Hello {0}, today is {1}",
+	"object _0", "object _1"
+	)]
+
+	[InlineData(
+	"Hello {name}, today is {date}",
+	"object name", "object date"
+	)]
+
+	[InlineData(
+	"{date} Hello {name}, today is {date}",
+	"object name", "object date"
+	)]
+
+	[InlineData(
+	"Hello {name:s}, today is {date:d}",
+	"string name", "DateOnly date"
+	)]
+
+	[InlineData(
+	"{i:int} {s:string}  {dt:DateTime}  {dto:DateTimeOffset}  {d:DateOnly}  {t:TimeOnly}  {0}		   {1}          {name}",
+	"int i", "string s", "DateTime dt", "DateTimeOffset dto", "DateOnly d", "TimeOnly t", "object _0", "object _1", "object name"
+	)]
+
+	[InlineData(
+	"{i:i}   {s:s}       {dt:dt}        {dto:dto}             {d:d}         {t:t}",
+	"int i", "string s", "DateTime dt", "DateTimeOffset dto", "DateOnly d", "TimeOnly t"
+	)]
+	public void Extracts_Parameters(string input, params string[] expected)
+	{
+		var sut = new ParameterBuilder(input);
+		var model = sut.Build(new(new("Ressource1.resx", input, 42)));
+		var writer = new ParameterDeclarationWriter(model);
+
+		var actual = writer.ToCSharp();
+
+		actual.Should().BeEquivalentTo(expected.ToCommaDelimited());
+	}
+
+	[Theory]
+	[InlineData(
+		"Hello {0}",
+		"object _0"
+	)]
+	[InlineData(
+		"Hello {name}",
+		"object name"
+	)]
+	[InlineData(
+		"Hello {0}, today is {1}",
+		"object _0, object _1"
+	)]
+	[InlineData(
+		"Hello {name}, today is {date}",
+		"object name, object date"
+	)]
+	[InlineData(
+		"{date} Hello {name}, today is {date}",
+		"object date, object name"
+	)]
+	public void Declares_Parameters_In_Signature(string input, string expected)
+	{
+		var sut = new ParameterBuilder(input);
+		var model = sut.Build(new(new("Ressource1.resx", input, 42)));
+		var writer = new ParameterDeclarationWriter(model);
+
+		var actual = writer.ToCSharp();
+
+		actual.Should().Be(expected);
+	}
+
+	[Theory]
+	[InlineData(
+		"Hello {0}",
+		"_0"
+	)]
+	[InlineData(
+		"Hello {name}",
+		"name"
+	)]
+	[InlineData(
+		"Hello {0}, today is {1}",
+		"_0, _1"
+	)]
+	[InlineData(
+		"Hello {name}, today is {date}",
+		"name, date"
+	)]
+	[InlineData(
+		"{date} Hello {name}, today is {date}",
+		"date, name"
+	)]
+	[InlineData(
+		"{0} {i:i} {s:s} {dt:dt} {dto:dto} {d:d} {t:t} {1}",
+		"_0, i, s, dt, dto, d, t, _1"
+	)]
+	public void Passes_Parameters_In_Invocation(string input, string expected)
+	{
+		var sut = new ParameterBuilder(input);
+		var parameters = sut.Build(new(new("Ressource1.resx", input, 42)));
+
+		var actual = parameters.Select(x => x.DisplayName).ToCommaDelimited();
+
+		actual.Should().BeEquivalentTo(expected);
 	}
 }
